requirements.txt
scapy==2.5.0
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
psutil==5.9.6
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
plotly==5.15.0
pyyaml==6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pymongo==4.5.0
redis==5.0.0
requests==2.31.0
cryptography==41.0.4
geoip2==4.7.0
maxminddb==2.2.0
